Refactor agent run service boundary
Move the low-level agent loop behind AgentRunService and split context, model, tool, finish, history, mutation, memory, and planning behavior into class-owned modules. Delete the old runAgent wrapper and root-level helper files so callers use the service boundary directly.
