+++ src/core/agent/constants.ts
@@ -1,0 +1,3 @@
+export const DEFAULT_MAX_STEPS = 100;
+export const STREAM_UPDATE_INTERVAL_MS = 75;
+export const INTERRUPTED_SUMMARY = 'Run interrupted by host request';
