+++ src/core/agent/context/index.ts
@@ -1,0 +1,6 @@
+export { AgentRunContextBuilder } from './run-context-builder.js';
+export type {
+  AgentRunContext,
+  BuildAgentRunContextArgs,
+  BuildInitialAgentMessagesArgs,
+} from './types.js';
