+++ src/core/agent/finish/index.ts
@@ -1,0 +1,8 @@
+export { AgentRunFinisher } from './run-finisher.js';
+export type {
+  AgentRunFinishedValue,
+  FinishAgentRunArgs,
+  FinishAgentRunLogging,
+  FinishAssistantResponseArgs,
+  MaybeFinishInterruptedArgs,
+} from './types.js';
