+++ src/core/agent/history/index.ts
@@ -1,0 +1,2 @@
+export { AgentHistorySanitizer } from './history-sanitizer.js';
+export type { AssistantWithTools, SanitizeAgentHistoryArgs, ToolMessage } from './types.js';
