+++ src/core/agent/index.ts
@@ -1,0 +1,2 @@
+export { AgentRunService } from './service.js';
+export type { RunAgentOptions } from './types.js';
