+++ src/core/agent/memory/index.ts
@@ -1,0 +1,6 @@
+export { AgentMemoryCheckpointTracker } from './memory-checkpoint-tracker.js';
+export type {
+  CreateAgentMemoryCheckpointStateArgs,
+  TrackAgentMemoryToolResultArgs,
+  AgentMemoryCheckpointState,
+} from './types.js';
