+++ src/core/agent/model/index.ts
@@ -1,0 +1,6 @@
+export { AgentModelTurnService } from './model-turn-service.js';
+export type {
+  AccumulateAgentUsageArgs,
+  AgentModelTurnResult,
+  RequestAgentModelTurnArgs,
+} from './types.js';
