+++ src/core/agent/mutation/types.ts
@@ -1,0 +1,11 @@
+import type { ToolCall } from '@/core/types.js';
+
+export type MutationState = {
+  executedMutationCommands: string[];
+};
+
+export type TrackAgentMutationResultArgs = {
+  state: MutationState;
+  effectiveCall: ToolCall;
+  result: { ok: boolean; output?: unknown };
+};
