+++ src/core/agent/planning/types.ts
@@ -1,0 +1,8 @@
+import type { PlanItem } from '@/core/tools/toolkits/internal/update-plan.js';
+import type { AgentPlanState } from '../types.js';
+
+export type ParseAgentPlanStateArgs = {
+  output: unknown;
+};
+
+export type { AgentPlanState, PlanItem };
