+++ src/core/agent/tools/index.ts
@@ -1,0 +1,8 @@
+export { AgentToolTurnService } from './tool-turn-service.js';
+export { AgentToolDispatcher } from './tool-dispatcher.js';
+export type {
+  AgentToolTurnResult,
+  ExecuteAgentToolTurnArgs,
+  HandleAgentToolResultArgs,
+  HandleAgentToolTurnArgs,
+} from './types.js';
