+++ src/core/agent/utils/abort.ts
@@ -1,0 +1,11 @@
+export function isAbortError(err: unknown): boolean {
+  if (!(err instanceof Error)) {
+    return false;
+  }
+
+  return (
+    err.name === 'AbortError' ||
+    err.name === 'APIUserAbortError' ||
+    /aborted/i.test(err.message)
+  );
+}
